Refactored turn tests a little
Moved restrictTurn to own class so that it can be used in multiple tests
